refactor with routers
